% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_error_function.R
\name{classification_error}
\alias{classification_error}
\title{Title Given a confusion matrix, return the classification error.}
\usage{
classification_error(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{The confusion matrix for prediction summary using the model.}
}
\value{
number of incorrectly classified individuals / total number of individuals
}
\description{
Title Given a confusion matrix, return the classification error.
}
\examples{
A <- matrix(c(12, 24, 25, 30))
classification_error(A)
# 0.8681319

classification_error(diag(c(12, 24)))
# 0

}
