% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_SNP_XGBoost_Object_function.R
\name{Create_Single_SNP_Object}
\alias{Create_Single_SNP_Object}
\title{For a single SNP, create an SNPFastImpute Object}
\usage{
Create_Single_SNP_Object(df, a, size)
}
\arguments{
\item{df}{the dataframe containing NAs, p columns of SNPs, n rows of samples.}

\item{a}{the column indicator of the SNP in the dataset.}

\item{size}{the windows size around the SNP to use as predictor variables.}
}
\value{
an object for the corresponding SNP.

If the SNP has missing value, then this object is a list with 6 elements:

1. model_fit: indicator whether we need to fit a model for the SNP.
2. SNP_position: position of the SNP.
3. NA_positions: position of the missing values.
4. train_xgboost: the xgboost object to train the model. 
5. pred_data: samples that is not missing for this SNP.
6. pred_label: an empty vector to store the future predicted labels. 
7. error: an empty number to store the future prediction error of the model.
}
\description{
For a single SNP, create an SNPFastImpute Object
}
\examples{
df <- readRDS("Test_df.rds")
single_SNP_Obj <- Create_Single_SNP_Object(df, 1, 10)
## return a list of SNP related information for model building.

Create_Single_SNP_Object(df, 50, 200)
## Error message. 
Stop as the size of the windows are larger than the number of SNPs in the dataframe. 

Create_Single_SNP_Object(matrix(NA, 10, 5), 1, 1)
## Should print a warning message that all samples for this SNP are NA's.
## return a list containing a model_fit value equal to false. 


}
