% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_SNP_XGBoost_Object_function.R
\name{Create_SNP_XGBoost_Object}
\alias{Create_SNP_XGBoost_Object}
\title{Input a dataframe, generate an object for the SNP missing value imputation with xgboost.}
\usage{
Create_SNP_XGBoost_Object(df, size = 10)
}
\arguments{
\item{df}{A dataframe of the SNPs which we need to impute missing values}

\item{size}{A windows size which we control for our imputation. Default value is 50.}
}
\value{
a list of list of 2 contents:  

1. NA_cols: column index with missing values  

2. Multiple_SNP_Object: corresponding SNP_Objects for these SNPs
}
\description{
Input a dataframe, generate an object for the SNP missing value imputation with xgboost.
}
\details{
This function takes in a matrix (p columns of SNPs, n rows of samples) with NAs and 
returns an object for our imputation function.
}
\examples{
df <- readRDS("Test_df.rds")
## This is a short matrix containing 100 SNPs with 20 samples. 

xgboost_snp_obj <- Create_SNP_XGBoost_Object(df)

xgboost_snp_obj <- Create_SNP_XGBoost_Object(df, size = 200)



}
