% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggplot.error_time}
\alias{ggplot.error_time}
\title{Imputation error and time plot.}
\usage{
ggplot.error_time(dta_orig, targetlist, sizelist = c(10))
}
\arguments{
\item{dta_orig}{The original}

\item{targetlist}{A list of target objects which we want to compare the imputation error rate and time}

\item{sizelist}{A list of windows sizes to be tested. Default value is only size of 10.}
}
\value{

}
\description{
Imputation error and time plot.
}
