% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute_Genotype_XGBoost_function.R
\name{Impute_GenoType_XGBoost}
\alias{Impute_GenoType_XGBoost}
\title{Impute missing SNPs for the input dataset.}
\usage{
Impute_GenoType_XGBoost(df, size = 10, num_class = 3, nrounds = 100)
}
\arguments{
\item{df}{The original dataset including the missing SNPs to be imputed.}

\item{size}{The windows size to use as the training dataset for each SNP, default: 10.}

\item{num_class}{Number of classes of response variable (types of SNPs), default: 3.

params A list of parameters for the xgboost model building. 
Default: nrounds = 100, booster = "gbtree", objective = "multi:softprob", 
num_class = 3, eval_metric = "mlogloss".}

\item{nrounds}{Number of fitting rounds, default: 100.}
}
\value{
The predicted missing genotypes.
}
\description{
Impute missing SNPs for the input dataset.
}
\details{
In our model, we try to use the types of SNPs around each missing SNP to predict the missing value. 
For each missing value, we need to use the size n of SNPs around it as predictors,
and use the non-missing samples for this SNP position as the training dataset.
}
\examples{
data("Test_df")
df <- Test_df
predict_df <- Impute_GenoType_XGBoost(df, size = 10)
## May take several seconds to finish.
## Should return a dataset where the missing values are filled by predicted values.  

}
\references{
Kabisch, Maria, Ute Hamann, and Justo Lorenzo Bermejo. "Imputation of 
missing genotypes within LD-blocks relying on the basic coalescent and 
beyond: consideration of population growth and structure." BMC genomics 18.1 (2017): 798.

Tianqi Chen and Carlos Guestrin, "XGBoost: A Scalable Tree Boosting System",
22nd SIGKDD Conference on Knowledge Discovery and Data Mining, 2016, \url{https://arxiv.org/abs/1603.02754}
}
