\name{SNPFastImpute-package}
\alias{SNPFastImpute}
\alias{SNPFastImpute}
\docType{package}
\title{
\packageTitle{SNPFastImpute}
}
\description{
\packageDescription{SNPFastImpute}
}
\details{
Main functions: 
classfication_error, Create_Single_SNP_Object, Impute_GenoType_XGBoost, Generate_NA
}
\author{
\packageAuthor{SNPFastImpute}

Maintainer: \packageMaintainer{SNPFastImpute}
}
\references{
Kabisch, Maria, Ute Hamann, and Justo Lorenzo Bermejo. "Imputation of 
missing genotypes within LD-blocks relying on the basic coalescent and 
beyond: consideration of population growth and structure." BMC genomics 18.1 (2017): 798.

Tianqi Chen and Carlos Guestrin, "XGBoost: A Scalable Tree Boosting System",
22nd SIGKDD Conference on Knowledge Discovery and Data Mining, 2016, \url{https://arxiv.org/abs/1603.02754}
}


\keyword{ package }
