% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2df.R
\name{vcf2df}
\alias{vcf2df}
\title{Read a vcf file, output the corresponding dataframe.}
\usage{
vcf2df(filename)
}
\arguments{
\item{filename}{The vcf file we want to read and generate the dataframe we need to 
generate the xgboost dataframe.}
}
\value{
A dataframe which we need to generate the xgboost data structure.
}
\description{
Read a vcf file, output the corresponding dataframe.
}
\examples{
filename <- "Test.vcf"
output_df <- vcf2df(filename)


}
