% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_error_function.R
\name{classification_error}
\alias{classification_error}
\title{Classification error calculator.}
\usage{
classification_error(df1, df2, NA_positions)
}
\arguments{
\item{df1}{The original dataset that contains only true missing values, not introduced 
missing values.}

\item{df2}{The second dataset that has all missing values filled in.}

\item{NA_positions}{The positions where we introduced NA in the original dataset.}
}
\value{
number of incorrectly classified individuals / total number of individuals
}
\description{
Classification error calculator.
}
\examples{
data("SNP_orig_sub")
data("SNP_NA_df02")
df_fill <- Impute_GenoType_XGBoost(SNP_NA_df02$SNP_NA_df)
NA_positions <- SNP_NA_df02$NP_generate_positions
classification_error(SNP_orig_sub, df_fill, NA_positions)

}
