% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_SNP_XGBoost_Object_function.R
\name{Create_SNP_XGBoost_Object}
\alias{Create_SNP_XGBoost_Object}
\title{Input a dataframe, generate an object for the SNP missing value imputation with xgboost.}
\usage{
Create_SNP_XGBoost_Object(df, size = 50)
}
\arguments{
\item{df}{A dataframe of the SNPs which we need to impute missing values}

\item{size}{A windows size which we control for our imputation. Default value is 50.}
}
\value{
a list of list of 3 contents:  
1. XGBoost_dataset  
1.1. xgb_train: X_train  
1.2. xgb_test: X_test  
1.3. test_df_label  
1.4. snp_NA  
1.5. pred: Empty now, to store the predicted values for each SNP  
1.6. error: Empty now, to store the prediction errors for each SNP  
1.7. xgb_pred  
1.8. SNP  
  
2. df_NA  
3. df_noNA
}
\description{
Input a dataframe, generate an object for the SNP missing value imputation with xgboost.
}
\examples{
df <- readRDS("Test_df.rds")
xgboost_snp_obj <- Create_SNP_XGBoost_Object(df)

xgboost_snp_obj <- Create_SNP_XGBoost_Object(df, size = 100)

}
