% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_NA.R
\name{NA_Generator}
\alias{NA_Generator}
\title{Function to introduce additional NAs for prediction error calculation.}
\usage{
NA_Generator(SNP_df, percent)
}
\arguments{
\item{SNP_df}{Original data of SNP. It may contain some SNP}

\item{percent}{The final percentage of total NA's we want in the dataset.}
}
\value{
a list of variables:  
1. SNP_NA_df: the dataset with additional SNPs.   
2. NA_percent_orig: original missing percentable in the dataset.   
3. NA_percent_generate: introduced NA percentage.   
4. NP_generate_positions: the positions of the introduced NA's.
}
\description{
Function to introduce additional NAs for prediction error calculation.
}
\examples{
SNP_df <- readRDS("data/SNP_orig_sub.rds")
## original NA ratio is 0.018
SNP_NA_df <- NA_Generator(SNP_df, 0.2)
## Introduced another 18\% of NAs. 
# saveRDS(SNP_NA_df, file = "data/SNP_NA_df_list.rds")
SNP_NA_df <- NA_Generator(SNP_df, 0.01)
## Should report an error: alread has 1.8\% NA's, higher than the 
## target percentage. 

}
